requirements.txt
Django==4.2.8
google-generativeai==0.3.1
python-dotenv==1.0.0
Pillow==10.1.0
